Implemented Shop Scenes / Changed Music Changing
Implemented the hardware and paint shops into the game.  Changed the way
the music singleton and scene transitions load new songs.  Now on the
fade object attached to the camera, the new song name can be written in
as a string.  If the scene loads and the current song playing is
different, the musichandler will change to the new specified song.